sql/pgwire: Avoid uint{16, 32} -> int{16, 32} overflows, which could cause panics

+++ sql/pgwire/encoding.go
@@ -129,20 +129,20 @@ func (b *readBuffer) getBytes(n int) ([]byte, error) {
 	return v, nil
 }
 
-func (b *readBuffer) getInt16() (int16, error) {
+func (b *readBuffer) getUint16() (uint16, error) {
 	if len(b.msg) < 2 {
 		return 0, errors.Errorf("insufficient data: %d", len(b.msg))
 	}
-	v := int16(binary.BigEndian.Uint16(b.msg[:2]))
+	v := binary.BigEndian.Uint16(b.msg[:2])
 	b.msg = b.msg[2:]
 	return v, nil
 }
 
-func (b *readBuffer) getInt32() (int32, error) {
+func (b *readBuffer) getUint32() (uint32, error) {
 	if len(b.msg) < 4 {
 		return 0, errors.Errorf("insufficient data: %d", len(b.msg))
 	}
-	v := int32(binary.BigEndian.Uint32(b.msg[:4]))
+	v := binary.BigEndian.Uint32(b.msg[:4])
 	b.msg = b.msg[4:]
 	return v, nil
 }
